Add composable environment for running local metacpan-api
If using plain docker-compose, running

    export COMPOSE_PROJECT_NAME=metacpan
    export COMPOSE_FILE="docker-compose.yml:docker-compose.localapi.yml"
    docker-compose up

Will set up a complete MetaCPAN environment with both the
frontend (http://localhost;5001) and backend (http://localhost:5000) as
well as ElasticSearch (http://localhost:9200) online.  This makes use
the previous `docker-compose.yml` configuration, _composing_ in further
settings for the local API to happen.

But that's a mouthful, so use `bin/metacpan-docker` to make all that
simple instead.

See the [compose CLI][0] [docs][1] for the gory details.

[0]: https://docs.docker.com/compose/reference/envvars/
[1]: https://docs.docker.com/compose/reference/overview/#use--f-to-specify-name-and-path-of-one-or-more-compose-files