Allow web service to start
The web service does not work on a initial start because it tries to
create `/metacpan-web/var` which is a read-only bind mount. My initial
thoughts were to override it with a docker-compose.override.yml, but I
think users should be able to quickly fire up docker without having to
write an override file. Thus I added a volume for `/metacpan-web/var`.
Another option is to create var in the metacpan directory with a .keep
file and commit that but ignore all other files in it.