fix metacpan_web config
The configuration being used was broken because it was a copy of the old
main config file, and hadn't been updated with the changes that have
been made. There shouldn't be any need to mirror these changes. Instead
we can just override the values we care about, specifically the URLs
used.