Remove api and web configs
Use the configs in the respective repos. I'm not sure why this was done,
but this setup makes it very easy for the configs to get out of sync.
Since we are not using docker compose in production, I don't think we
should make the config file setup any more complicated it needs to be.